refactor: 나의 보따리 불러오기 API 문서화 및 서비스 패키지 변경
